Remove duplicate worker instantiation
It seems like the worker instantiated and saved to the variable "worker" is the one used to create the service, however, it is not the one added to the local list of workers. An unused NEW worker instance is instead added to the list. Effectively, each loop we create two worker instances, one in vain.